Add `cluster` feature
Closes #1110
