+++ src/graphql/client.rs
@@ -1,1 +1,2 @@
+pub mod cluster;
 pub mod derives;
